requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
groq==0.4.1
sentence-transformers==2.7.0
chromadb==0.4.24
rank-bm25==0.2.2
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.0
PyPDF2==3.0.1
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.0.0
pdf2image==1.16.3
numpy==1.26.4
tokenizers==0.19.1